+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.1.1, 2024-02-09:
+- Module code improvement to pass Shopware extension validator.
+- [smartform] Fix payment cancellation and payment error for Smartform.
+
 # 4.1.0, 2024-01-24:
 - [smartform] Smartform integration.
 - Set return mode to POST by default.